- implement the global aperture of cyclotron by 4 arguments in cyclotron element: 		MINR, MAXR, MINZ, MAXZ. - if a particle is out of cyclotron aperture or out of the field map,    it will be deleted after this step and its parameters are recorded in the *.loss file - change data type of particle id from int to size_t - many minor cosmetics of display massage

+++ classic/5.0/src/AbsBeamline/BeamBeam.cpp
@@ -48,11 +48,11 @@ void BeamBeam::accept(BeamlineVisitor &visitor) const {
     visitor.visitBeamBeam(*this);
 }
 
-bool BeamBeam::apply(const int &i, const double &t, double E[], double B[]) {
+bool BeamBeam::apply(const size_t &i, const double &t, double E[], double B[]) {
     return false;
 }
 
-bool BeamBeam::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool BeamBeam::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     return false;
 }
