+++ classic/5.0/src/AbsBeamline/Corrector.h
@@ -77,9 +77,9 @@ public:
     /// Return the plane on which the corrector acts.
     virtual Plane getPlane() const = 0;
 
-    virtual bool apply(const int &i, const double &t, double E[], double B[]);
+    virtual bool apply(const size_t &i, const double &t, double E[], double B[]);
 
-    virtual bool apply(const int &i, const double &t, Vector_t &E, Vector_t &B);
+    virtual bool apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B);
 
     virtual bool apply(const Vector_t &R, const Vector_t &centroid, const double &t, Vector_t &E, Vector_t &B);
