+++ classic/5.0/src/AbsBeamline/Lambertson.cpp
@@ -48,11 +48,11 @@ void Lambertson::accept(BeamlineVisitor &visitor) const {
     visitor.visitLambertson(*this);
 }
 
-bool Lambertson::apply(const int &i, const double &t, double E[], double B[]) {
+bool Lambertson::apply(const size_t &i, const double &t, double E[], double B[]) {
     return false;
 }
 
-bool Lambertson::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool Lambertson::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     return false;
 }
