+++ classic/5.0/src/AbsBeamline/Solenoid.cpp
@@ -159,7 +159,7 @@ void Solenoid::addKT(int i, double t, Vector_t &K) {
     K += temp;
 }
 
-bool Solenoid::apply(const int &i, const double &t, double E[], double B[]) {
+bool Solenoid::apply(const size_t &i, const double &t, double E[], double B[]) {
     Vector_t Ev(0, 0, 0), Bv(0, 0, 0);
     Vector_t Rt(RefPartBunch_m->getX(i), RefPartBunch_m->getY(i), RefPartBunch_m->getZ(i));
     if(apply(Rt, Vector_t(0.0), t, Ev, Bv)) return true;
@@ -174,7 +174,7 @@ bool Solenoid::apply(const int &i, const double &t, double E[], double B[]) {
     return false;
 }
 
-bool Solenoid::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool Solenoid::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     Vector_t tmpE(0.0, 0.0, 0.0), tmpB(0.0, 0.0, 0.0);
     const Vector_t tmpR(RefPartBunch_m->getX(i) - dx_m, RefPartBunch_m->getY(i) - dy_m, RefPartBunch_m->getZ(i) - startField_m - ds_m);
