+++ classic/5.0/src/AbsBeamline/Solenoid.h
@@ -55,9 +55,9 @@ public:
 
     virtual void addKT(int i, double t, Vector_t &K);
 
-    virtual bool apply(const int &i, const double &t, double E[], double B[]);
+    virtual bool apply(const size_t &i, const double &t, double E[], double B[]);
 
-    virtual bool apply(const int &i, const double &t, Vector_t &E, Vector_t &B);
+    virtual bool apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B);
 
     virtual bool apply(const Vector_t &R, const Vector_t &centroid, const double &t, Vector_t &E, Vector_t &B);
