+++ classic/5.0/src/AbsBeamline/TravelingWave.cpp
@@ -272,7 +272,7 @@ void TravelingWave::addKT(int i, double t, Vector_t &K) {
 */
 
 
-bool TravelingWave::apply(const int &i, const double &t, double E[], double B[]) {
+bool TravelingWave::apply(const size_t &i, const double &t, double E[], double B[]) {
     Vector_t Ev(0, 0, 0), Bv(0, 0, 0);
     Vector_t Rt(RefPartBunch_m->getX(i), RefPartBunch_m->getY(i), RefPartBunch_m->getZ(i));
     if(apply(Rt, Vector_t(0.0), t, Ev, Bv)) return true;
@@ -287,7 +287,7 @@ bool TravelingWave::apply(const int &i, const double &t, double E[], double B[])
     return false;
 }
 
-bool TravelingWave::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool TravelingWave::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     double tmpcos, tmpsin;
 
     Vector_t tmpR(RefPartBunch_m->getX(i) - dx_m, RefPartBunch_m->getY(i) - dy_m , RefPartBunch_m->getZ(i) - startField_m - ds_m);
